media-indexer/backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "media-indexer-backend"
version = "2.0.0"
description = "FastAPI backend for MKLanLocal v2.0"
requires-python = ">=3.12"
dependencies = [
  "alembic>=1.14.0",
  "argon2-cffi>=23.1.0",
  "apscheduler>=3.10.4",
  "fastapi>=0.115.0",
  "huggingface_hub>=0.32.0",
  "httpx>=0.28.1",
  "itsdangerous>=2.2.0",
  "onnxruntime>=1.20.0",
  "orjson>=3.10.0",
  "Pillow>=11.0.0",
  "pgvector>=0.3.5",
  "psutil>=7.0.0",
  "psycopg[binary]>=3.2.0",
  "pydantic-settings>=2.6.0",
  "python-multipart>=0.0.12",
  "sqlalchemy>=2.0.36",
  "torch>=2.5.0",
  "transformers>=4.46.0",
  "uvicorn[standard]>=0.32.0",
  "sse-starlette>=1.6.5",
  "pytesseract>=0.3.13",
  "scikit-learn>=1.5.0",
  "numpy>=1.26.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
